rename files and pack unused code into folder
